added prompt modal in it's own component
